weaviate_complete_project 2/backend/requirements.txt
flask
flask-cors
requests
beautifulsoup4
sentence-transformers
weaviate-client
faiss-cpu
numpy
tqdm
transformers
torch
